requirements.txt
flask==1.1.1
flask-mongoengine==0.9.5
python-decouple==3.1
pandas>=0.23.3,<1
pymongo>=3.10.1
numpy>=1.14.5,<2
scipy>=1.1.0,<2
scikit-image>=0.14.1,<1
jupyter>=1.0.0,<2
nbformat>=4.4.0,<5
requests>=2.21.0
boto3>=1.9.0
xarray==0.12.1
netCDF4==1.5.3
